animation system improvements

+++ enc_temp_folder/acd9ba6bc55a459085e31a5d39ea3/Player.cpp
@@ -236,8 +236,8 @@ void Player::_handleAction(int actionCode) {
 
 void Player::_updateRotation() {
 	//Crosshair position relative to player
-	double x = _crosshairX - _x;
-	double y = _crosshairY - _y;
+	int x = _crosshairX - _x;
+	int y = _crosshairY - _y;
 
 	//Distance from player to crosshair
 	double r = sqrt(x*x + y*y);
